requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
decorator==5.1.1
executing==0.9.1
Faker==13.15.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.23.1
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rich==12.5.1
shellingham==1.5.0
six==1.16.0
stack-data==0.3.0
tqdm==4.64.0
traitlets==5.3.0
typer==0.6.1
urllib3==1.26.11
validators==0.18.2
wcwidth==0.2.5
weaviate-client==3.7.0
